% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Results2017}
\alias{Results2017}
\title{Results from French 2017 Presidential Election}
\format{A data frame with 461,721 rows and 11 variables:
\describe{
 \item{Region}{Region Code (INSEE)}
 \item{RegionNom}{Region Name}
 \item{Departement}{Departement Code (INSEE)}
 \item{DepartmentNom}{Department Name}
 \item{CommuneCode}{Commune Code (elections)}
 \item{CommuneName}{Commune Name (elections)}
 \item{CommuneCodeINSEE}{Commune Code (INSEE)}
 \item{CommuneNomINSEE}{Commune Name (INSEE)}
 \item{Round}{Round 1/2}
 \item{Candidate}{Name of candidate}
 \item{Votes}{Number of Votes}
}}
\usage{
Results2017
}
\description{
Data frame of results from both rounds of the 2017 presidential election in France.
}
\keyword{datasets}
